restService/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.1
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.5
whitenoise==5.2.0